Restructure curl function
It is possible to reuse an established connection with curl.
This speeds up the transfer around 4 times up.
The problems are that a Segmentation fault occurs
when we pass the curl variable into different functions using a pointer.
That's why The code is moved into connection.c.
This is not the best solution but it is better than nothing.